changed the default value

+++ src/gazeArbiter/include/iCub/gazeArbiterModule.h
@@ -183,7 +183,8 @@
  * 24/04/11 : added fixed neck pitch option                                                         @author Rea
  * 28/03/11 : added the limits for the allowed fixation points in 3d space                          @author Rea
  * 29/03/11 : added new parameters to the config file and embedded previously added ones            @author Rea
- * 06/04/11 : added a new port for acquiring the template and sending it to the objectsColl.       @author Rea
+ * 06/04/11 : added a new port for acquiring the template and sending it to the objectsColl.        @author Rea
+ * 11/04/11 : introduced boolean flag for allowing refinement of the final position                 @author Rea 
  */
 
 
@@ -208,9 +209,9 @@ class gazeArbiterModule:public yarp::os::RFModule {
     std::string handlerPortName;                // name of the handler port (comunication with respond function)
     std::string configFile;                     // configuration file of cameras
     int ratethread;                             // time constant for ratethread
-    int xoffset;                                // offset for the 3D point along x
-    int yoffset;                                // offset for the 3D point along y
-    int zoffset;                                // offset for the 3D point along z
+    double xoffset;                                // offset for the 3D point along x
+    double yoffset;                                // offset for the 3D point along y
+    double zoffset;                                // offset for the 3D point along z
     double xmax, xmin;                          // limits for the allowed fixation point (x axis)
     double ymax, ymin;                          // limits for the allowed fixation point (y axis)
     double zmax, zmin;                          // limits for the allowed fixation point (z axis)
